xserver: add xdm user with role access to system_r and xdm_r
Sync with upstream's xserver changes.

Previously reverted in dada9b3defc6c44e73d56adf245a5812c3f08404. The
reasoning for the revert:
```
This commit added the sddm user to the xserver module.
This caused problems loading the xserver module if the user did not
exist on the system.
```
no longer applies, as upstream hit this issue here:
https://github.com/SELinuxProject/refpolicy/issues/488 and resolved it.

Fixes: https://github.com/gentoo/hardened-refpolicy/issues/7

Signed-off-by: Rahul Sandhu <rahul@sandhuservices.dev>
Closes: https://github.com/gentoo/hardened-refpolicy/pull/8
Signed-off-by: Kenton Groombridge <concord@gentoo.org>