init: add tunable to allow mounton selinux config
By default the kanidm-unixd systemd unit uses ProtectSystem=strict which
runs the service in a read-only mount namespace. However, the SELinux
runtime features require the SELinux config to be writable to create
user home dir equivalency rules in the policy. Therefore init requires
being able to mount on selinux_config_t dirs to create a read-writeable
bind-mount inside the namespace.

Add a tunable for this functionality.

Signed-off-by: Kenton Groombridge <concord@gentoo.org>
Signed-off-by: Jason Zaman <perfinion@gentoo.org>