files: add a default file context spec for /proc
There are many situations where having a default label for /proc is
helpful in early boot, generally before /proc is mounted in cases such
as plymouth. Having a /proc label also doesn't really hurt here, and
makes labelling more robust before the system is fully initalised.

As proc_t is defined in the kernel policy module, and not the files
policy module, move the labelling for /proc into the kernel policy
module so we can use proc_t for it.

Another example of this being useful is for mounting namespaces, such
as used by systemd for auditd:

[   19.902620] audit: type=1400 audit(1663630933.439:3): avc:  denied  { mounton } for  pid=1062 comm="(auditd)" path="/run/systemd/unit-root/proc" dev="dm-3" ino=67581 scontext=system_u:system_r:init_t:s0 tcontext=system_u:object_r:unlabeled_t:s0 tclass=dir permissive=1

Closes: https://bugs.gentoo.org/871966
Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>
Signed-off-by: Jason Zaman <perfinion@gentoo.org>