incus: rm explicit fcontext for /usr/libexec/incus(/.*)?
When creating a VM, incusd needs to read an incus-agent binary
on the host to integrate it into the VM image.
If the agents delivered by incus are being stored inside
/usr/libexec/incus/agents/ on the host the default file
context of bin_t is sufficient whereas with container_engine_exec_t
another allow rule would be required to search and read dirs
and files in the agents/ sub-directory.

Signed-off-by: Marc Schiffbauer <mschiff@gentoo.org>
Signed-off-by: Jason Zaman <perfinion@gentoo.org>